requirements.txt
altgraph==0.15
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
macholib==1.9
MarkupSafe==1.0
pefile==2017.11.5
psutil==5.4.5
pycrypto==2.6.1
PyInstaller==3.3.1
python-dotenv==0.8.2
requests==2.18.4
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
